BR001: refactoring, bug fix, change type for price to BigDecimal

+++ src/main/java/org/warehouse/app/dto/product/ProductDto.java
@@ -3,13 +3,15 @@ package org.warehouse.app.dto.product;
 import lombok.Data;
 import org.warehouse.app.entity.Product;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductDto {
 
     private String article;
     private String name;
-    private Long lastIncomePrice;
-    private Long lastSalePrice;
+    private BigDecimal lastIncomePrice;
+    private BigDecimal lastSalePrice;
 
     public static ProductDto create(Product product) {
         ProductDto productDto = new ProductDto();
